requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
firebase-admin>=6.2.0
python-dotenv>=1.0.0
